it's moving (in your dreams)

+++ CommonModule/FieldState.h
@@ -3,7 +3,7 @@
 extern "C" {
 	class FieldState {
 	public:
-		uchar reserved = 0;
+		uint reserved = 0;
 		uint stateSize;
 
 		//group 4  together
